rework exception handling tests
