+++ tests/EndpointFixture.php
@@ -72,20 +72,4 @@ class EndpointFixture implements Interfaces\EndpointInterface
             ->will(new ReturnStub(json_encode($input->asArray())));
         return $mock;
     }
-
-    public function handleException(Throwable $e): Response
-    {
-        /** @var Response | \PHPUnit\Framework\MockObject\MockObject */
-        $mock = (new Generator())
-            ->getMock(Response::class);
-        $code = $e->getCode();
-        if ($code < 200 || $code > 599) {
-            $code = self::STATUS_ERROR; // Artificial test value
-        }
-        $mock->method('getStatusCode')
-            ->will(new ReturnStub($code));
-        $mock->method('getBody')
-            ->will(new ReturnStub($e)); // This is incorrect, but makes debugging tests easier
-        return $mock;
-    }
 }
